+++ package.json
@@ -2,10 +2,10 @@
   "dependencies": {
     "axios": "^0.23.0",
     "chess.js": "^0.12.0",
+    "cron": "^1.8.2",
     "express": "^4.17.1",
     "google-auth-library": "^7.10.0",
     "mongodb": "^4.1.3",
-    "node-cron": "^3.0.0",
     "serve-favicon": "^2.5.0"
   }
 }
